Move code to Utils, drop IDs identical to source
