+++ pontoon/machinery/tests/test_utils.py
@@ -5,11 +5,18 @@ import pytest
 from django.core.cache import cache
 from django.core.management import call_command
 
+from pontoon.base.models import Resource
 from pontoon.machinery.utils import (
     get_google_automl_translation,
+    get_llm_string_id,
     get_machinery_service_cache_key,
 )
-from pontoon.test.factories import LocaleFactory
+from pontoon.test.factories import (
+    EntityFactory,
+    LocaleFactory,
+    ResourceFactory,
+    SectionFactory,
+)
 
 
 @pytest.fixture
@@ -94,3 +101,64 @@ def test_warmup_automl_command_bypasses_cache(automl_project_id, automl_locale):
 
     # The warmup made a real request even though a cached value was present.
     assert client.translate_text.call_count == 1
+
+
+@pytest.mark.django_db
+@pytest.mark.parametrize(
+    "format, section_key, entity_key, expected",
+    [
+        # Formats without a section key
+        (Resource.Format.PROPERTIES, [], ["open-browser"], "open-browser"),
+        # Formats with the file name or section name as the section key
+        (
+            Resource.Format.XLIFF,
+            ["Client/Client.strings"],
+            ["Client/Client.strings", "Settings.Title"],
+            "Settings.Title",
+        ),
+        (
+            Resource.Format.XCODE,
+            ["Localizable.strings"],
+            ["Localizable.strings", "Settings.Title"],
+            "Settings.Title",
+        ),
+        (Resource.Format.INI, ["Strings"], ["Strings", "label"], "label"),
+        (Resource.Format.ANDROID, ["!ENTITY"], ["!ENTITY", "brandName"], "brandName"),
+        # Parts identical to the source string are dropped: the id in XLIFF and
+        # XCODE files, or the message id in gettext, optionally followed by the
+        # message context
+        (
+            Resource.Format.XCODE,
+            ["Localizable.strings"],
+            ["Localizable.strings", "Open browser"],
+            None,
+        ),
+        (Resource.Format.GETTEXT, [], ["Open browser"], None),
+        (Resource.Format.GETTEXT, [], ["Open browser", "toolbar"], "toolbar"),
+        # Nested keys are joined
+        (
+            Resource.Format.PLAIN_JSON,
+            [],
+            ["toolbar", "open-browser"],
+            "toolbar.open-browser",
+        ),
+        # Missing key
+        (Resource.Format.PROPERTIES, [], [], None),
+    ],
+)
+def test_get_llm_string_id(format, section_key, entity_key, expected):
+    resource = ResourceFactory(format=format)
+    section = SectionFactory(key=section_key, resource=resource)
+    entity = EntityFactory(
+        key=entity_key, string="Open browser", resource=resource, section=section
+    )
+
+    assert get_llm_string_id(entity) == expected
+
+
+@pytest.mark.django_db
+def test_get_llm_string_id_without_section():
+    resource = ResourceFactory(format=Resource.Format.PROPERTIES)
+    entity = EntityFactory(key=["open-browser"], resource=resource, section=None)
+
+    assert get_llm_string_id(entity) == "open-browser"
